requirements.txt
fastapi
python-jose[cryptography]
passlib[bcrypt]
python-multipart
uvicorn
sqlalchemy